Add new mayTrap MachineInstr property
This can be used to avoid hoisting potentially) trapping instructions out
of conditional branches that may not be taken.
This patch introduces three new instruction flags to model the common
trapping behaviour of CHERI-MIPS and CHERI-RISC-V instructions:
 - mayTrapOnSealedInput: This instruction traps if input capabilities
   are sealed (e.g. CIncOffset)
 - mayTrapOnUntaggedInput: This instruction traps if input capabilities
   are untagged (e.g. CSetBounds/CAndPerm)
 - mayTrap: This instruction may trap under certain conditions. This can
   be used for CHERI instructions that have stricter requirements.

The new MachineInstr::mayTrap() that returns true if
  a) The instruction is marked as mayTrapOnSealedInput and one of the
     use operands could potentially be sealed.
  b) The instruction is marked as mayTrapOnUntaggedInput and one of the
     use operands could potentially be untagged.
  c) The instruction is marked as mayTrap.

In order to mark instructions as mayTrapOnSealedInput/mayTrapOnUntaggedInput,
the TableGen Constraint string for machine instructions has been extended
to add two new flags: "@traps_if_sealed $ropnd" and
"@traps_if_untagged $opnd". If TableGen sees this flag for one of the
operands it will mark the instruction as having the mayTrapOnSealedInput
and/or mayTrapOnUntaggedInput flag (see the ConstraintCheckingCap.td test)
so these instruction flags should not be set explicitly.

To determine the value for mayTrap() without being extremely
conservative, MachineOperand has a new maybeSealed()/maybeUntagged()
accessor that knows that e.g. FrameIndex are always valid tagged and
unsealed capabilities.
In preparation for future optimizations this patch also adds two more
instruction flags:
 - defMayBeSealed: The output of this instruction can be a sealed
   capability (defaults to false and should be set for CSeal, etc). It
   also implicitly set for capability loads.
 - defIsAlwaysTagged: The output of this instruction is always a tagged
     capability (e.g. CSetBounds, CSealEntry).
The flags will be used to infer whether a MachineOperand could be sealed
or untagged (e.g. all results of capability modification instructions
are always unsealed, so uses of that def can be ignored for
mayTrapOnSealedInput constraints.
We may also want to add new TII hooks to perform more detailed analysis:
(e.g. we know that a CIncOffsets generated for FrameIndices is not
only unsealed but also always tagged[1]).

This is not ideal for code generation but required to prevent
MachineLICM from hoisting of instructions that can trap before a
conditional check.
I initially used the simpler approach of marking CHERI instructions with
hasSideEffects=1, however this regresses code generation too much since
it prevents *any* kind of reordering (such as moving instrctions into
delay slots). This is overly restrictive since we only need to prevent
LLVM from hoisting instructions that may trap from basic blocks that
might not be entered into ones that are always traversed.

MachineLICM has been updated to not hoist instructions where mayTrap()
returns true and that are not guaranteed to execute.
Once the constraints are added for RISC-V and MIPS, this wil fix
https://github.com/CTSRD-CHERI/llvm-project/issues/348 and
https://github.com/CTSRD-CHERI/llvm-project/issues/359

I am doing this in a separate patch so that can easily be reverted in case
we decide to change to tag-clearing instead of trapping behaviour in the
future.

[1] This is not strictly true since the offset could be so large that
    the result becomes unrepresentable but in that case the stack has
    already overflowed so this is unlikely to produce traps that may
    not occur on all paths through the function.