Infer whether MachineOperands can be untagged
We can currently infer this for instructions marked as
MCID::DefIsAlwaysTagged, and in the future we could do more sophisticated
analyses (e.g. in-bounds incoffset on a stack slot) but that's probably not
worth the effort. This allows for more hoisting of instructions out
of loops (e.g. users of CSetBounds/CAndPerms instructions)