Add a no_preserve_cheri_tags function attribute
This is the opposite of the existing must_preserve_cheri_tags and can be
used to indicated that a memcpy/memmove will never copy data and can
therefore be inlined even if it larger than the capability size and aligned
to less than the capability size.
The motivating case for this is that we currently call memcpy when
performing structure assignment for struct { long a; long b;} rather than
using two loads+stores.