Add a test that the str*->memcpy optimizations set no_preserve_tags
All the string functions in SimplifyLibcalls.cpp should mark the created
memcpy instruction as non-tag-preserving to allow inlining. Testing strcpy
has exposed another bug in SimplifyLibcalls: strcpy is not being optimized
for purecap code since TargetLibraryInfoImpl::isValidProtoForLibFunc()
only treats strcpy as a builtin function if the argument is i8* (in
address space 0). This will be fixed in a follow-up commit.