Avoid memcpy() calls for cap-size copies of scalar types
For example, memcpy((long*)a, (long*)b, CAP_SIZE); can now be inlined
as two loads and two stores rather than having to call memcpy. Previously,
we conservatively assumed that any copy > CAP_SIZE could contain tags.

The only exception are void *, (unsigned) char * and std::byte * since
those are used to point to buffers with arbitrary data (possibly including
capabilities). This is not strictly required for unsigned char, but we
believe it's common for code to use uint8_t as a type for arbitrary
buffers.