Conservatively set no_preserve_tags for MemberExprs
If we have an expression such as __builtin_memcpy(buf, &s.not_a_cap, len);
and we can see the declaration for s and the entire VarDecl for s does not
contain tags, then we can set no_preserve_tags. However, we can't set the
attribute if the struct does contain capabilities (even though not_a_cap
does not not contain capabilities), since the memcpy() could extend beyond
the field. We also need to consider fields before the address of since we
could be copying a array and extending past the end of the initial struct.