[CHERI StdCompat] Emit a Clang error for unsupported builtins
Previously the compiler would just fail with a cannot select/invalid
instruction predicates error. With this commit we now emit a C frontend
error pointing at the line where the unsupported builtin is used instead
of crashing. The end result is essentially the same but the new
behaviour is a lot more user-friendly.