[InstSimplify] Add folds for @llvm.cheri.cap.high.get
As part of this refactor some of the CHERI get intrinsics folding code.
This refactoring introduces a minor change to getoffset folding as well:
we no longer fold arbitrary getoffset(setoffset(A, B)) -> B since the
setoffset could have changed the bounds interpretation and therefore
also the result. This optimization is unlikely to make a difference in
practise and is somewhat incorrect so I believe dropping it makes sense.