[CHERI-RISC-V] Restrict CToPtr usage to ISAv8 compatibility
CToPtr will be removed from the CHERI-RISC-V ISA in the future and will
not be part of a standardized version. We emit a branch-less form of
`x.tag ? x.addr - auth.base : 0` when we see the intrinsic in non-ISAv8
compilation mode instead.

When expanding `ptrtoint` in hybrid mode we also no longer subtract the
base of DDC as part of the expansion, since DDC relocation is no longer
part of xcherimin, and subtracting the base would result in an integer
that points to a different address when dereferenced.