[CHERI-RISC-V] Introduce a +xcheri-no-v8-compat target feature
When set, this feature allows the compiler to assume new capability
semantics (tag-clearing instead of trapping and removal of DDC/PCC
relocation). This will allow hoisting previously trapping instructions
out of loops which cannot be done unconditionally now as we still want
to support building and running the last CheriBSD release with the latest
dev branch of the compiler. We therefore need an opt-in flag that can be
set by the CheriBSD build system.