[CHERI-RISC-V] Restrict CFromPtr usage to ISAv8 compatibility
We emit `x == 0 ? null : CSetOffset(auth, x)` when we see the intrinsic in
non-ISAv8 compilation mode instead. When expanding `inttoptr` in hybrid
mode, we do not include the base of DDC as part of the expansion (since
DDC relocation is no longer part of ISAv9). Adding the base would result
in a capability that points to a different address when dereferenced.
Therefore, inttoptr is expanded to `x == 0 ? null : CSetAddr(DDC, x)`.
It would be nice if we could just use `CSetAddr(DDC, x)` unconditionally
for non-constant arguments, but that could result in semantic differences
(tagged vs untagged) for integers that happen to be zero at run time.