[UPSTREAM] Fix m_PtrToIntSameSize for CHERI
We have to compare the integral range of the pointer not the raw bitwidth.
This fixes the 64-bit memcpy-preserve-tags-assume-aligned.ll tests and with
a change to use i32 ptrtoint for the RISC-V 32-bit test that one also
continues to work correctly.

The fix will need to be upstreamed with an appropriate LangRef change.